Move PreprocessorHelper out of ClusteringPreprocessor
